+++ apps/sim/app/(landing)/comparisons/comparison-sections.ts
