+++ apps/sim/app/(landing)/comparisons/components/brand-icon-tile/index.ts
