+++ apps/sim/app/(landing)/comparisons/components/comparison-cards/index.ts
