+++ apps/sim/app/(landing)/comparisons/components/comparison-table/index.ts
