+++ apps/sim/app/(landing)/comparisons/components/fact-value/index.ts
