+++ apps/sim/app/(landing)/comparisons/utils.ts
