+++ apps/sim/app/sitemap.ts
@@ -9,7 +9,7 @@ import {
   ALL_COMPETITORS,
   getLatestVerifiedDate,
   SIM_LATEST_VERIFIED,
-} from '@/app/(landing)/comparison/utils'
+} from '@/app/(landing)/comparisons/utils'
 import { ALL_CATALOG_MODELS, MODEL_PROVIDERS_WITH_CATALOGS } from '@/app/(landing)/models/utils'
 
 /** One sitemap entry per author, timestamped by their most recently updated post. */
@@ -172,9 +172,9 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       : SIM_LATEST_VERIFIED
 
   const comparisonPages: MetadataRoute.Sitemap = [
-    { url: `${baseUrl}/comparison`, lastModified: comparisonLastModified },
+    { url: `${baseUrl}/comparisons`, lastModified: comparisonLastModified },
     ...ALL_COMPETITORS.map((competitor) => ({
-      url: `${baseUrl}/comparison/${competitor.id}`,
+      url: `${baseUrl}/comparisons/${competitor.id}`,
       lastModified: competitorLastModified(competitor),
     })),
   ]
